==> 01_download_GBIF_data_exercise.Rmd <==
---
title: "Downloading geographic occurrence data from GBIF"
output: 
  html_document:
    theme: readable
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(fig.width=12, fig.height=8, 
                      echo=TRUE, warning=FALSE, message=FALSE,
                      tidy = TRUE, collapse = TRUE)
```

## Background
The public availability of species distribution data has increased substantially in the last 10 years: occurrence information, mostly in the form of geographic coordinate records for species across the tree of life, representing hundreds of years of biological collection effort are now available. The global Biodiversity Information Facility (www.gbif.org) is one of the largest data providers, hosting more than one billion records (Sept 2018) from a large variety of sources.

## Outcomes
After this exercise you will be able to retrieve species occurrence information from GBIF from within R. You will be equipt with example data from your group of interest for the follow upcoming exercises. See https://ropensci.org/tutorials/rgbif_tutorial.html for a more exhaustive tutorial on the rgbif package.

## Exercise
We will use the rgbif package to obtain occurrence records from GBIF. You can find the relevant functions for each task in the parentheses. You can get help on each function by typing `?FUNCTIONNAME`. 


1. Familiarize yourself with the `rgbif` package and download the occurrence data for one of the species we saw in the field. (`name_suggest`, `occ_Search`)
2. Look at the data and a quick plot (`head`, `plot`).
3. Now find out how many occurrences for flowering plants are availble from Brazil.  (`name_suggest`, `occ_count`)
4. Download all records for flowering plants from Rio Grande do Norte (more or less). Make sure to keep an eye on the `limit` argument. (`occ_search`)
5. Save the downloaded data as .txt or .csv to the working directory. (`write_csv`, `write_delim`)
<!-- 6. DOwnload records from at least one additional databases, for example https://www.idigbio.org/ and http://www.splink.org.br/. -->


<!-- 1. Download data for a single species in your group of interest and save them in a data.frame. For this exercise make sure to limit the download to a reasonable number of records (e.g. 5000), using the `limit` argument (`name_suggest`, `occ_search`). -->
<!-- 2. Look at the data and a quick plot (`head`, `plot`). -->
<!-- 3. How many records are there for your entire group of interest, globally and in Brazil (`occ_count`)? -->
<!-- 4. Download all data for your group from Brazil, or if the group is very large from the North East. -->
<!-- 5. Save the downloaded data as .txt or .csv to the working directory. (`write_csv`, `write_delim`) -->
<!-- 6. You can check additional databases for more records of your group, for plants for example https://www.idigbio.org/ and http://www.splink.org.br/. -->

Occurrence records from the Global Biodiversity Information Facility can be downloaded from their web portal (www.gbif.org), but there are also several ways to obtain records from within R. Getting records from within R has the advantage, that the analyses are fully reproducible and can be repeated automatically. rgbif is a package that offers exhaustive functionalities for obtaining data from GBIF. Here we will focus on the basic functions to obtain data for the follow up exercises. You can learn more on rgbif with [this vignette](https://ropensci.org/tutorials/rgbif_tutorial.html). Please note that there are licence requirements if you use GBIF data, which include a proper citation including DOI ([learn more here](https://www.gbif.org/citation-guidelines)).

## Setup
In this exercise we will use the rgbif library for communication with GBIF and the tidyverse library for data management.

```{r setup, include=FALSE}
library(rgbif)
library(tidyverse)
```
