==> 03_sampling_biases_exercise.Rmd <==
---
title: "Quantifying sampling bias"
output: 
  html_document:
    theme: readable
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(fig.width=12, fig.height=8, 
                      echo=TRUE, warning=FALSE, message=FALSE,
                      tidy = TRUE, collapse = TRUE,
                      results = 'hold')
```


## Background
Sampling effort is a critical measure for the reliability of present/absence projections from species occurrence data. However, dataset from publicly available databases based on biological collections are often comprised from various sources and compiled over large time periods and information on sampling effort is therefore most of the time unavailable. Physical accessibility of a region has been identified as major predictor of sampling effort, but this effect might vary among datasets. The SampBias package allows quantifying the effect of different anthropogenic structures (roads, airports, cities) on sampling in any given dataset based on geographic gazetteers. You can find a description of the methods here and tutorials on the use of SampBias [here](https://github.com/azizka/sampbias/wiki). **NOTE SampBias is only available as beta version**

## Objectives
After this exercise you will be able to 
* quantify the effect of accessibility on the sampling pattern in species occurrence dataset
* have an idea on political factors biasing data collection

## Exercise
Helpful functions for answering each question are given in the brackets. In case you want to get a feeling for the functionality of SampBias, without using are, you can find a GUI app here.

1. Load the distribution data. (`read.csv`)
2. Run sampbias with the default settings. (`SamplingBias`)
3. Look at the run summary and visualize the results. How informative are the results? (`par, plot`)
4. Explore ?SamplingBias and try to change the relevant arguments to improve the results. Summarize and visualize again.
5. Explore the [Bio-Dem app](http://bio-dem.surge.sh/) to explore the effect of political factos on data collections in your group. 

## Possible questions for your research project
* How biased is your collection dataset by accesability?
* Which inffrastructure is most biasing?


## Library setup
You will need the following R libraries for this exercise, just copy the code chunk into you R console to load them. You might need to install some of them separately.

```{r}
library(tidyverse)
library(sampbias)
```
