==> 04_species_richness_maps_exercise.Rmd <==
---
title: "Species richness maps and species ranges"
output: 
  html_document:
    theme: readable
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(fig.width=12, fig.height=8, 
                      echo=TRUE, warning=FALSE, message=FALSE,
                      tidy = TRUE, collapse = TRUE,
                      results = 'hold')
```

## Background
Patterns of species richness and occurrence density are relevant information for study design, to communicate study results or to check data quality in any multi-species occurrence dataset. Species distributions can be estimated with a large array of methods, but in some cases a rough and quick estimation of species ranges is more suitable. The extent of occurrence based on convex hulls is used in conservation and macroecology to approximate species' ranges.

## Objectives
After this exercise you will be able to:
1. Visualize species richness patterns and occurrence record density in a quick and efficient way.
2. Estimate species distribution ranges based, and create polygons of species distributions.

## Exercise
1. Load the example data, and visualize the coordinates on a map.
2. Visualize the number of occurrence records and the number of species in a regular-sized grid (`RichnessGrid`).
3. Estimate the range for each species using a convex hull. Write the range maps to the working directory as ESRI-shape files (`writeOGR`, `plot`, `CalcRange`).
4. Visualize species richness per grid cell based on the range polygons (`RangeRichness`).
5. Visulaize species richenss in an equal area raster (`raster`, `spTransform`, `coordinates`).
6. Write the result rasters to a file on your disk

## Possible questions for your project
* Where is the diversity centre of the group?
* Where are centres of record collection?

## Library setup
You will need the following R libraries for this exercise, just copy the code chunk into you R console to load them. You might need to install some of them separately.

```{r}
library(tidyverse)
library(speciesgeocodeR)
library(raster)
library(rgdal)
```
